% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBurstSzFn.R
\name{calcBurstSzFn}
\alias{calcBurstSzFn}
\title{Calculate burst size function}
\usage{
calcBurstSzFn(in.v, in.nthr = 100, in.sc = 1/2)
}
\arguments{
\item{in.v}{Vector with intervals}

\item{in.nthr}{Total number of thresholds}

\item{in.sc}{Scaling factor for threshold interval. Provide as 1/x to obtain the maximum threshold size = x.}
}
\value{
Data frame with threshold lengths and the value of burst size function.
}
\description{
Calculates burst size function \code{Phi(tau)} according to Eq.4 of
Dobrzynski & Bruggeman (2009) Elongation dynamics shape bursty transcription and translation. PNAS, 106(8)
Formula: \code{Phi (tau) = total no. of intervals / number of intervals longer than tau}
}
\details{
The input should contain a single-column data with intervals between events such as
firings of a neuron, or production of a protein. The function searches for the number of events
longer than a threshold. The procedure is repeated for a series of threshold lengths.
The output data frame contains two columns with the number of events for each of the thresholds.
}
\examples{
# generate a series of exponentially distributed intervals
df = rexp(1000, 1)
phi=calcBurstSzFn(v$V1, 100, 1/2)
plot(phi$t, phi$b, type = 'l')

}
