% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCI.R
\name{calcCI}
\alias{calcCI}
\title{Calculate 95\% CI from the mean and SD (normal assumption)}
\usage{
calcCI(in.dt, in.col.mn, in.col.sd, in.col.n, in.clevel = 0.95)
}
\arguments{
\item{in.dt}{Input data table}

\item{in.col.mn}{Name of the column containing the mean}

\item{in.col.sd}{Name of the column containing SD}

\item{in.col.n}{Name of the column containing sample size}

\item{in.clevel}{Confidence level, default 95\%}
}
\value{
Data table with additional columns with lower and upper CI bounds
}
\description{
From: http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)
}
