% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTrajCI.R
\name{calcTrajCI}
\alias{calcTrajCI}
\title{Calculate the mean and CI around time series}
\usage{
calcTrajCI(in.dt, in.col.meas, in.col.by = NULL, in.type = c("normal",
  "boot"), ...)
}
\arguments{
\item{in.dt}{Data table in long format}

\item{in.col.meas}{Name of the column with the measurement}

\item{in.col.by}{Column names for grouping (default NULL - no grouping). Typically, you want to use at least a column with time.}

\item{in.type}{Choice of normal approximation or boot-strapping}

\item{...}{Other params passed to smean.cl.normal and smean.cl.boot; these include \code{conf.int} for the confidence level, \code{B} for the number of boot-strapping iterations.}
}
\value{
Datatable with columns: Mean, lower and upper CI, and grouping columns if provided.
}
\description{
Calculate the mean and CI around time series
}
\examples{

require(tca)

# generate synthetic time series; 100 time points long, with 10 randomly placed NAs
dt.tmp = genTraj(100, 10, 6, 3, in.addna = 10)

# calculate single stats from all time points
calcTrajCI(dt.tmp, 'objCyto_Intensity_MeanIntensity_imErkCor')

# calculate the mean and CI along the time course
calcTrajCI(dt.tmp, 'objCyto_Intensity_MeanIntensity_imErkCor', 'Metadata_RealTime')
}
