% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPar.R
\name{checkDigits}
\alias{checkDigits}
\title{Check whether a string consists only from digits}
\usage{
checkDigits(x)
}
\arguments{
\item{x}{Input string}
}
\value{
True if the input string consists only from digits, False otherwise.
}
\description{
Check whether a string consists only from digits
}
\examples{
checkDigits('1111')
checkDigits('1111cccc')
}
