% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPar.R
\name{checkLogical}
\alias{checkLogical}
\title{Check whether a string matches TRUE/FALSE, T/F, or T.../F...}
\usage{
checkLogical(x)
}
\arguments{
\item{x}{Input string}
}
\value{
True if the input string matches the pattern, False otherwise.
}
\description{
Check whether a string matches TRUE/FALSE, T/F, or T.../F...
}
\examples{
checkLogical('TRUE')
checkLogical('xxxTxxx')
}
