% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseNrandom.R
\name{chooseNrandom}
\alias{chooseNrandom}
\title{Randomly choose rows by group}
\usage{
chooseNrandom(in.dt, in.n, in.col.id, in.col.group)
}
\arguments{
\item{in.dt}{Input data table}

\item{in.n}{Number of rows to sample}

\item{in.col.id}{Name of the column with unique row id's}

\item{in.col.group}{Name of the grouping column}
}
\value{
Data table with sampled rows
}
\description{
Takes N random rows from a data table. If grouping column provided, sampling is performed by group.
}
