% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPar.R
\name{convertStringList2Types}
\alias{convertStringList2Types}
\title{Converts string elements of a named list to apporpriate types}
\usage{
convertStringList2Types(in.l)
}
\arguments{
\item{in.l}{Named list fo strings.}
}
\value{
Named list with elements converted to appropriate types.
}
\description{
Strings that consist of digits are converted to type \code{numeric}, strings with TRUE/FALSE, T/F, or T.../F... to \code{logical}.
}
\examples{
 l.tst = list()
 l.tst$aaa = '1000'
 l.tst$bbb = '1000xx'
 l.tst$ccc = 'True'
 l.tst$ddd = 'xxxTrue'
 l.res = convertStringListToTypes(l.tst)
 str(l.res)
}
