% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadCSV2lineHeader.R
\name{freadCSV2lineHeader}
\alias{freadCSV2lineHeader}
\title{Read a CSV file with a 2-line header}
\usage{
freadCSV2lineHeader(in.file, in.col.rem = NULL)
}
\arguments{
\item{in.file}{Name of the CSV file.}

\item{in.col.rem}{String vector with column names to remove}
}
\value{
Data table with column names set according to 2 first lines of the header, i.e. merged with '_'.
}
\description{
A CSV file with a 2-line header is frequently output by CellProfiler when merging separate object outputs is set
in ExportToSpreadsheet module. The output contains merged outputs from different objects (e.g. cytosol, nucleus),
which means that the header consists of 2 lines, first with object name, second with the measurement name.
These two lines are merged in this function.
}
