% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTraj.R
\name{genTraj}
\alias{genTraj}
\title{Generate synthetic CellProfiler output with single cell time series}
\usage{
genTraj(in.ntpts = 60, in.ntracks = 10, in.nfov = 6, in.nwells = 1,
  in.addna = NULL)
}
\arguments{
\item{in.ntpts}{Number of time points (default 60)}

\item{in.ntracks}{Number of tracks per FOV (default 10)}

\item{in.nfov}{Number of FOV (default 6)}

\item{in.nwells}{Number of wells (default 1)}

\item{in.addna}{Number of NAs to add randomly in the data (default NULL)}
}
\value{
Data table with the follwoing columns: Metadata_Site, Metadata_Well, Metadata_RealTime, objCyto_Intensity_MeanIntensity_imErkCor (normal distributed),
objNuc_Intensity_MeanIntensity_imErkCor (normal distributed), objNuc_Location_X and objNuc_Location_Y (uniform ditributed), TrackLabel
}
\description{
Generate synthetic CellProfiler output with single cell time series
}
