% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotTheme.R
\name{ggplotTheme}
\alias{ggplotTheme}
\title{Custom ggPlot theme based on theme_bw}
\usage{
ggplotTheme(in.font.base = 12, in.font.axis.text = 12,
  in.font.axis.title = 12, in.font.strip = 14, in.font.legend = 12)
}
\arguments{
\item{in.font.legend}{}
}
\description{
Custom ggPlot theme based on theme_bw
}
\examples{

}
