% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normTraj.R
\name{normTraj}
\alias{normTraj}
\title{Normalize Trajectory}
\usage{
normTraj(in.dt, in.meas.col, in.rt.col = "RealTime", in.rt.min = 10,
  in.rt.max = 20, in.by.cols = NULL, in.robust = TRUE,
  in.type = "z.score")
}
\arguments{
\item{in.dt}{Data table in long format}

\item{in.meas.col}{String with the column name to normalize}

\item{in.rt.col}{String with the colum name holding time}

\item{in.rt.min}{Lower bound for time period used for normalization}

\item{in.rt.max}{Upper bound for time period used for normalization}

\item{in.by.cols}{String vector with 'by' columns to calculate normalization per group; if NULL, no grouping is done}

\item{in.robust}{Whether robust measures should be used (median instead of mean, mad instead of sd); default TRUE}

\item{in.type}{Type of normalization: z.score or mean (i.e. fold change w.r.t. mean); default 'z-score'}
}
\value{
Returns original dt with an additional column with normalized quantity.
}
\description{
Returns original dt with an additional column with normalized quantity.
The column to be normalised is given by 'in.meas.col'.
The name of additional column is the same as in.meas.col but with ".norm" suffix added.
Normalisation is based on part of the trajectory;
this is defined by in.rt.min and max, and the column with time in.rt.col.#'
}
