% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraj.R
\name{plotTraj}
\alias{plotTraj}
\title{Plot individual time series}
\usage{
plotTraj(dt.arg, x.arg, y.arg, group.arg, facet.arg = NULL,
  summary.arg = c("none", "mean", "meanCInorm", "meanCIboot"),
  facet.ncol.arg = 2, line.col.arg = NULL, xlab.arg = "Time",
  ylab.arg = "Fl. int.", plotlab.arg = "", dt.stim.arg = NULL, stim.x.arg,
  stim.y.arg, maxrt.arg = 60, xaxisbreaks.arg = 10, xlim.arg = NULL,
  ylim.arg = NULL)
}
\arguments{
\item{dt.arg}{Data table in long format}

\item{x.arg}{String with column name of x-axis data}

\item{y.arg}{String with column name of y-axis data}

\item{group.arg}{String with column name of the grouping variable for trajectories. This is usually a trackID.}

\item{facet.arg}{String with oclumn name of the variable used for plot facets (optional)}

\item{summary.arg}{Optional summary stats to plot on top of individual time series. Choose from \code{c('none', 'mean', 'meanCInorm', 'meanCIboot')},
to plot no extra stats, the mean, the mean with 95\% CI calculated from the normal distribution approximation, the mean with 95\% CI calculated via bootstrapping, respectively.}

\item{facet.ncol.arg}{Number of facetting columns}

\item{line.col.arg}{String with variable name to use for time series colouring (default NULL)}

\item{xlab.arg}{String with x-axis label}

\item{ylab.arg}{String with y-axis label}

\item{plotlab.arg}{String with plot title}

\item{dt.stim.arg}{Data table with stimulation curve}

\item{stim.x.arg}{String with column name of x-axis stimulation data}

\item{stim.y.arg}{String with column name of y-axis stimulation data}

\item{maxrt.arg}{Max value of x-axis label}

\item{xaxisbreaks.arg}{Value of x-axis interval}

\item{xlim.arg}{Limit of x-axis}

\item{ylim.arg}{Limit of y-axis}
}
\value{
ggplot2 plot object
}
\description{
Plot individual time series
}
\examples{

require(tca)

# generate synthetic time series
dt = genTraj()

# plot each group in a separate facet; add population mean
plotTraj(dt.arg = dt, x.arg = 'Metadata_RealTime', y.arg = 'objNuc_Intensity_MeanIntensity_imErkCor', group.arg = 'TrackLabel', facet.arg = 'Metadata_Site', summary.arg = 'mean')
}
