% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrajRibbon.R
\name{plotTrajRibbon}
\alias{plotTrajRibbon}
\title{Plot time series with CI ribbon}
\usage{
plotTrajRibbon(dt.arg, x.arg, y.arg, group.arg = NULL, col.arg = NULL,
  dt.stim.arg = NULL, stim.x.arg, stim.y.arg, ribbon.lohi.arg = c("Lower",
  "Upper"), ribbon.fill.arg = "grey50", ribbon.alpha.arg = 0.5,
  xlab.arg = NULL, ylab.arg = NULL, plotlab.arg = NULL,
  legendpos.arg = c("top", "right", "bottom"))
}
\arguments{
\item{dt.arg}{Data table in long format with mean time series and upper and lower bound for the CI}

\item{x.arg}{String with column name of x-axis data (default: Metadata_RealTime)}

\item{y.arg}{String with column name of y-axis data (default: Mean)}

\item{group.arg}{String with column name of the grouping variable for trajectories (default: Metadata_Well).}

\item{col.arg}{Vector with string elements with colour palette for time series curves (default: NULL).}

\item{dt.stim.arg}{Data table with stimulation time series (default: NULL).}

\item{stim.x.arg}{String with column name of x-axis of dt.stim.arg.}

\item{stim.y.arg}{String with column name of y-axis of dt.stim.arg.}

\item{xlab.arg}{String with x-axis label}

\item{ylab.arg}{String with y-axis label}

\item{plotlab.arg}{String with plot title}
}
\value{
ggplot2 plot object
}
\description{
A ribbon with confidence interval for the mean is plotted around the group mean.
Optionally an additional curve is plotted, e.g. a stimulation time course.
}
\examples{

require(tca)

# generate synthetic time series
dt = genTraj()

# clculate 90\% CI
dt.aggr = calcTrajCI(in.dt = dt, in.col.meas = 'objNuc_Intensity_MeanIntensity_imErkCor', in.col.by = 'Metadata_Site', in.type = 'normal', conf.int = 0.9)

# plot each group in a separate facet; add population mean
plotTrajRibbon(dt.arg = dt.aggr, x.arg = 'Metadata_RealTime', y.arg = 'Mean')
}
