% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putItemFirst.R
\name{putItemFirst}
\alias{putItemFirst}
\title{Put a factor first}
\usage{
putItemFirst(in.dt, in.col, in.item)
}
\arguments{
\item{in.dt}{Input data table}

\item{in.col}{Name of the column to work on}

\item{in.item}{String with a factor to put as first}
}
\value{
Function changes state of the column provided by in.dt
}
\description{
Take a column with levels and put a specified factor first
}
