% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPar.R
\name{readPar}
\alias{readPar}
\title{Return a list with parameter names and their values read from xlsx or csv file}
\usage{
readPar(in.fname, in.sheet.idx = 1)
}
\arguments{
\item{in.fname}{Name of the xlsx file.}

\item{in.sheet.idx}{Integer with the sheet number in the xlsx to process.}
}
\value{
Named list with parameters and their values.
}
\description{
The xlsx or csv file has to contain at least two columns:
1st column with parameter names
2nd column with parameter values
}
