% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepNsignif.R
\name{signif_df}
\alias{signif_df}
\title{Keep a number of significant digits in a data.frame.}
\usage{
signif_df(df, digits)
}
\arguments{
\item{df}{Data in data.frame format}

\item{digits}{Number of significant digits to keep}
}
\value{
Returns original data frame with numeric fields trimmed to n significant digits.
}
\description{
Keep a number of significant digits in a data.frame.
}
\examples{

}
