% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepNsignif.R
\name{signif_dt}
\alias{signif_dt}
\title{Keep a number of significant digits in a data.table}
\usage{
signif_dt(dt, digits)
}
\arguments{
\item{dt}{Data in data.table format}

\item{digits}{Number of significant digits to keep}
}
\value{
Returns original data table with numeric fields trimmed to n significant digits.
}
\description{
Keep a number of significant digits in a data.table
}
