% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecShift.R
\name{vecShift}
\alias{vecShift}
\title{Shift vector values to right or left
Modified from: https://stackoverflow.com/a/37069049/1898713}
\usage{
vecShift(in.x, in.n, in.invert = FALSE, in.fill = NA, in.circle = F)
}
\arguments{
\item{in.x}{Vector for which to shift values}

\item{in.n}{Number of places to be shifted.
Positive numbers will shift to the right by default.
Negative numbers will shift to the left by default.
The direction can be inverted by the invert parameter.}

\item{in.invert}{Whether or not the default shift directions
should be inverted.}

\item{in.fill}{The value that should be inserted by default.}

\item{in.circle}{If TRUE, inserted values will be taken from shifted vector}
}
\value{
Shifted vector
}
\description{
Shift vector values to right or left
Modified from: https://stackoverflow.com/a/37069049/1898713
}
\examples{
v = 1:10

# Shift by two elements right, fill with NAs
vecShift(v, 2)

# Shift by two elements right, fill with 0's
vecShift(v, 2, in.fill = 0)

# Shift by two elements left, fill circular
vecShift(v, -2, in.circle = T)

}
